refactor isolate feature mathematics
